gateway/requirements.txt
# Web Framework
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# HTTP Client
httpx==0.27.2

# Authentication
pyjwt==2.9.0
cryptography==43.0.3

# AWS/Cloudflare R2
boto3==1.35.50

# Database (Convex client)
requests==2.32.3

# CORS
python-multipart==0.0.17

# Logging
python-json-logger==3.1.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Development
black==24.10.0
ruff==0.7.3
